// src/app/models/Pokemon.ts
export class Pokemon {
  id: number;
  name: string;
  type1: string;
  type2: string;
  weight: string;
  height: string;
  stats: {} [];
  types: {} [];
  sprites: {};
  imageDefault: string;
  imageShiny: string;
  url: string;
  ability1: string;
  ability2: string;
  ability3: string;
  attack: string;
  defense: string;
  hp: string;
  specialAttack: string;
  specialDefense: string;
  speed: string;
  results: {} [];
}